Quickstart with docker-compose (#25)
Using Docker Compose to bring up the MLX-API server and
the MLX-UI Dashboard, along with the MySQL database and
Minio S3 storage backend.
The catalog gets preloaded with sample Components, Datasets, Models
and Notebooks which can be browsed. The asset metadata and sample 
code can be downloaded and new assets can be registered.
Sample pipeline code can be generated for each asset type, however
their execution on a Kubeflow Pipelines cluster is not enabled.