Add login for mlx-ui (#41)
Add authentication in `/login` path. Use basic auth and simple
user model from a configuration in local file system. After a
successfull login, a cookie will be set in the browser and the user
info will be store in server. Currently, memorystore is used to
store session information. we can move it to persistent storage
later.

Signed-off-by: Yihong Wang <yh.wang@ibm.com>