Fix disable execution setting and enable rate limiting (#52)
* Fix disable run setting and enable rate limiting
Signed-off-by: Andrew-Butler <Andrew.Butler@ibm.com>

* Only show settings link for admin role (#48)

Read the cookie value to distinguish a user's role and
only show the settings link to users with admin role.

Signed-off-by: Yihong Wang <yh.wang@ibm.com>

* Add login for mlx-ui (#41)

Add authentication in `/login` path. Use basic auth and simple
user model from a configuration in local file system. After a
successfull login, a cookie will be set in the browser and the user
info will be store in server. Currently, memorystore is used to
store session information. we can move it to persistent storage
later.

Signed-off-by: Yihong Wang <yh.wang@ibm.com>

* Explicitly specify which routes to rate limit
Signed-off-by: Andrew-Butler <Andrew.Butler@ibm.com>

Co-authored-by: Yihong Wang <yh.wang@ibm.com>