[mlx-ui] Add js lib from kubeflow dashboard (#117)
In order to use ml-pipeline-ui in the iframe, it depends on a JS lib
from kubeflow dashboard on `/dashboard_lib.bundle.js` path.
Directly copy that JS lib from kubeflow centraldashboard to mlx-ui
as one of the static files. In case, mlx-ui is bound to `/` path,
it can serve the JS lib for ml-pipeline-ui in the iframe.

Signed-off-by: Yihong Wang <yh.wang@ibm.com>