Disable UI cache in Quickstart
With Docker Compose, the API and UI run locally and
the API has caching for GET requests. So there is no
real advantage in having yet another layer of caching
in the UI. In fact, the UI cache is often getting in
the way of testing the catalog by adding or deleting
assets.

Signed-off-by: Christian Kadner <ckadner@us.ibm.com>